# Ammo performance improvements