… dist/aframe-physics-system.js