Basic set of Ammo perf improvements, as per #15.

One slight hesitation about this is that people who have used previous versions of aframe-physics-system, and disregarded the rules about not moving static entities could have bugs exposed by moving to this version.

When we publish the version containing tis fix, we should add a note to the release note about that.

I *don't* think it's worth preserving the old behaviour with a flag, because the old behaviour is bugged, and I don't want to have to preserve bugged behaviour forever.