@diarmidmackenzie 2022-11-19
On my PC, I'm seeing median tick "before" perf reduce from ~1.5msecs to ~0.5msecs with these fixes.

Overall median tick reduces from ~3.5msecs to ~2.5 msecs.