@kylebakerio 2022-11-20
nice!

we could leave in some kind of `.error()` check for an attempt to move a static body being caught, with an explanation that they should change it to a kinematic body?