@kylebakerio 2022-11-20
Weird, why don't I see the `syncToPhysics()` function body here? Was it added previously?