@diarmidmackenzie 2022-11-21
> Weird, why don't I see the `syncToPhysics()` function body here? Was it added previously?

Not sure where you are looking... 

It's here:
https://github.com/c-frame/aframe-physics-system/blob/78290ee312bf0b7616942ec966203e3b42670c18/src/components/body/ammo-body.js#L373