@diarmidmackenzie 2022-11-21
> we could leave in some kind of `.error()` check for an attempt to move a static body being caught, with an explanation that they should change it to a kinematic body?

I don't see how to do this efficiently.

Inefficient options I can see are:
- check latest position vs. previous position every tick.  But that means checking every static object every tick.
- listen for the componentchanged event on a static object, in case the position component is updated.  But (a) that won't detect position changes made direct to the Object3D (which is the recommended way to update position) and (b) it requires code that executes every time any component on any static object changes.